feat: return  loan and rate book
Features:
- Avaliar livro [frontend e backend]
- Formulário de Avaliação
- Avaliação por estrelas
- Atualizando status de empréstimo ao devolver livro
- hide return button after return

+++ backend/src/main/java/com/unp/bibliotecavirtual/dto/request/LivroRequestDTO.java
@@ -36,6 +36,6 @@ public record LivroRequestDTO(
 
         @PositiveOrZero(message = "A avaliação não pode ser negativa")
         @Max(value = 10, message = "A avaliação máxima permitida é 10")
-        Integer avaliacao
+        Long avaliacao
 ) {
 }
